requirements.txt
Flask==3.0.3
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
Flask_PyMongo==2.3.0
PyJWT==2.9.0
pymongo==4.7.3
